pyproject.toml
[project]
name = "udon-peripheral-animation-tools"
version = "0.1.0"
description = "Stage-based image pipeline tooling for nice!view assets"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "pillow>=11.3.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["tools*"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
